feat: support ascii encoding
For better or for worse, sometimes you want to covert between a
Uint8Array and a string where each character of the string is
created from the character code of the corresponding byte from the
array.

Adds an `ascii` encoding which is similar to `binary` as used by
node buffers.

It's not called `binary` as people confuse this with `base2`.