fix: always return uint8arrays
3.1.0 incorporated internal refactors to use node `Buffer`s where
possible to increase performance when `alloc`ing new byte arrays.

It then returns those `Buffer`s but the problem is `Buffer` is not
completely compatible with `Uint8Array` as some methods with the
same name behave differently.

We can convert a `Buffer` to a `Uint8Array` without copying it
by using the 3-arg `Uint8Array` constructor so we should do that to
retain the performance characteristics of `Buffer` when `alloc`ing
but the compatibility of returning vanilla `Uint8Array`s.